requirements.txt
Flask==3.0.2
pandas==2.2.1
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.1
geopandas==0.14.3
shapely==2.0.4
pycountry==23.12.11 
folium
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27